requirements.txt
# Dependencies for the ThingSpeak + MQTT simulator
requests>=2.31.0
paho-mqtt>=1.6.1
